:recycle: Move FilterBar to components/shared
